refactor(views): views in one directory
