+++ client/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MainView from './src/components/MainView';
+import MainView from './src/components/Views/MainView';
 import { ConnectionContextProvider } from './src/ConnectionContext';
 
 export default function App() {
